scripts/knowledge-ingestion/requirements.txt
# Google Drive Knowledge Base Ingestion Dependencies

# Google Drive API
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# PDF Processing
marker-pdf==0.2.12
pymupdf==1.23.8
pypdf==3.17.1

# Embedding Models
openai==1.3.5
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1

# Database
supabase==2.0.2
psycopg2-binary==2.9.9
pgvector==0.2.4

# Async Processing
asyncio-mqtt==0.13.0
aiohttp==3.9.1
aiofiles==23.2.0

# Data Processing
numpy==1.24.4
pandas==2.1.3
scipy==1.11.4

# Configuration and Environment
pydantic==2.5.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and Monitoring
structlog==23.2.0
colorama==0.4.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.88.1

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1